[bugfix] fixed s2p_test.py and s2p_mosaic.py
using read_config_file and adding read_tiles
